requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+git://github.com/mischuh/p3common.git#egg=p3common
argh==0.26.2              # via watchdog
avro-python3==1.8.2
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
colorlog==3.1.2
decorator==4.2.1
idna==2.6                 # via requests
numpy==1.15.0             # via pandas
pandas==0.23.4
pathtools==0.1.2          # via watchdog
psycopg2-binary==2.7.5
python-dateutil==2.6.1    # via pandas
pytz==2017.3              # via pandas
pyyaml==3.13              # via watchdog
requests==2.18.4
six==1.11.0               # via python-dateutil
urllib3==1.22             # via requests
watchdog==0.8.3
